requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# created-by: conda 24.9.2
arpack=3.9.1=nompi_hf81eadf_101
brotli-python=1.1.0=py310h53e7c6a_2
bzip2=1.0.8=hfdf4475_7
ca-certificates=2024.12.14=h8857fd0_0
certifi=2024.12.14=pyhd8ed1ab_0
cffi=1.17.1=py310hfce808e_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
glpk=5.0=h3cb5acd_0
gmp=6.3.0=hf036a51_2
h2=4.1.0=pyhd8ed1ab_1
hpack=4.0.0=pyhd8ed1ab_1
hyperframe=6.0.1=pyhd8ed1ab_1
idna=3.10=pyhd8ed1ab_1
libblas=3.9.0=26_osx64_openblas
libcblas=3.9.0=26_osx64_openblas
libcxx=19.1.6=hf95d169_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libiconv=1.17=hd75f5a5_2
liblapack=3.9.0=26_osx64_openblas
liblzma=5.6.3=hd471939_1
libopenblas=0.3.28=openmp_hbf64a52_1
libsqlite=3.47.2=hdb6dae5_0
libxml2=2.13.5=he8ee3e7_1
libzlib=1.3.1=hd23fc13_2
llvm-openmp=19.1.6=ha54dae1_0
ncurses=6.5=hf036a51_1
numpy=2.2.1=py310h2874b4e_0
openssl=3.4.0=hd471939_0
pip=24.3.1=pyh8b19718_2
pycparser=2.22=pyh29332c3_1
pysocks=1.7.1=pyha55dd90_7
python=3.10.16=h5acdff8_1_cpython
python-igraph=0.11.8=py310h9558dc6_0
python_abi=3.10=5_cp310
readline=8.2=h9e318b2_1
requests=2.32.3=pyhd8ed1ab_1
setuptools=75.6.0=pyhff2d567_1
texttable=1.7.0=pyhd8ed1ab_1
tk=8.6.13=h1abcd95_1
tzdata=2024b=hc8b5060_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
zstandard=0.23.0=py310h41d873f_1
zstd=1.5.6=h915ae27_0
